requirements-dev.txt
gitpython>=3.1.30